# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rol/{id}": {
      "get": {
        "operationId": "RolController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rol"
        ]
      },
      "put": {
        "operationId": "RolController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRolDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rol"
        ]
      },
      "delete": {
        "operationId": "RolController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Rol"
        ]
      }
    },
...
    "/api/producto/linea/{id}": {
      "get": {
        "description": "Devuelve la línea asociada al producto indicado.",
        "operationId": "ProductoController_geLineaDelProducto",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID del producto.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Línea asociada al producto.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener la línea de un producto",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
...
    "/api/linea": {
      "post": {
        "description": "Registra una nueva línea en el sistema.",
        "operationId": "LineaController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLineaDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Línea creada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una línea",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
    "/api/linea/search-by": {
      "get": {
        "description": "Devuelve las líneas filtradas por denominación, con paginación.",
        "operationId": "LineaController_findByDenominacionFiltered",
        "parameters": [
          {
            "name": "denominacion",
            "required": false,
            "in": "query",
            "description": "Denominación a filtrar",
            "schema": {
              "example": "tornillo",
              "type": "string"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Cantidad de elementos a omitir",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "description": "Cantidad de elementos a retornar",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Líneas que coinciden con los filtros indicados.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ListadoConTotalDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/LineaDto"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Buscar líneas con filtros",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
    "/api/linea/select": {
      "get": {
        "description": "Devuelve las líneas activas que coinciden parcialmente por denominación.",
        "operationId": "LineaController_searchForSelection",
        "parameters": [
          {
            "name": "denominacion",
            "required": false,
            "in": "query",
            "description": "Denominación a filtrar.",
            "schema": {
              "example": "tornillo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Líneas activas que coinciden parcialmente por denominación.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SelectOption"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar líneas para selección",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
    "/api/linea/{id}": {
      "get": {
        "description": "Devuelve el detalle de la línea indicada.",
        "operationId": "LineaController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la línea.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Línea encontrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LineaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener una línea por ID",
        "tags": [
          "Gestion Productos"
        ]
      },
      "put": {
        "description": "Actualiza los datos de la línea indicada.",
        "operationId": "LineaController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la línea.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLineaDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Línea actualizada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una línea",
        "tags": [
          "Gestion Productos"
        ]
      },
      "delete": {
        "description": "Elimina de forma lógica la línea indicada.",
        "operationId": "LineaController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la línea.",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "usuarioId",
            "required": true,
            "in": "query",
            "description": "ID del usuario que realiza la eliminación.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Línea eliminada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MensajeDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una línea",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
    "/api/linea/find-all-for-superlinea/select": {
      "get": {
        "description": "Devuelve las superlíneas filtradas por denominación para asociarlas a una línea.",
        "operationId": "LineaController_findAllForSuperLineas",
        "parameters": [
          {
            "name": "denominacion",
            "required": false,
            "in": "query",
            "description": "Denominación a filtrar.",
            "schema": {
              "example": "tornillo",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Superlíneas filtradas por denominación.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SuperLineaDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar superlíneas para selección",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
    "/api/linea/{id}/audit": {
      "get": {
        "description": "Devuelve la información de auditoría de la línea indicada.",
        "operationId": "LineaController_findByIdConAuditoria",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "ID de la línea.",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Informacion de auditoria",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditoriaDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener la auditoría de una línea",
        "tags": [
          "Gestion Productos"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gestión Base - Distribuidora",
    "description": "La descripción de las  API  de la distribuidora",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Token JWT de acceso. Formato: Bearer <token>"
      }
    },
    "schemas": {
...
      "MensajeDto": {
        "type": "object",
        "properties": {
          "mensaje": {
            "type": "string",
            "description": "Mensaje para el front"
          }
        }
      },
      "ListadoConTotalDto": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Listado de resultados",
            "type": "array",
            "items": {
              "type": "array"
            }
          },
          "total": {
            "type": "number",
            "description": "Cantidad total de elementos"
          }
        },
        "required": [
          "data",
          "total"
        ]
      },
...
      "AuditoriaDto": {
        "type": "object",
        "properties": {}
      },
...
      "UpdateRolDto": {
        "type": "object",
        "properties": {}
      },
...
      "ReferenciaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Identificador único de la entidad referenciada.",
            "example": 1
          },
          "denominacion": {
            "type": "string",
            "description": "Denominación de la entidad referenciada.",
            "example": "Iveco"
          }
        },
        "required": [
          "id",
          "denominacion"
        ]
      },
      "LineaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 123,
            "description": "ID del la linea"
          },
          "denominacion": {
            "type": "string",
            "example": "tornillos",
            "description": "Denominación o nombre dela linea"
          },
          "stockMinimo": {
            "type": "number",
            "description": "Stock mínimo de la línea.",
            "example": 5
          },
          "utilizaStockMinimo": {
            "type": "boolean",
            "example": false,
            "description": "Indica si la línea utiliza stock mínimo."
          },
          "observacion": {
            "type": "string",
            "example": "",
            "description": "Observaciones varias sobre la linea"
          },
          "sistema": {
            "type": "number",
            "example": 1,
            "description": "de sistema no se puede editar ni eliminar"
          },
          "deletedAt": {
            "type": "string",
            "example": null,
            "description": "Fecha de eliminación (null si está activa)",
            "nullable": true
          },
          "superLinea": {
            "description": "Superlínea a la que pertenece la línea.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReferenciaDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "denominacion",
          "utilizaStockMinimo",
          "observacion",
          "sistema",
          "deletedAt",
          "superLinea"
        ]
      },
...
      "CreateLineaDto": {
        "type": "object",
        "properties": {
          "denominacion": {
            "type": "string",
            "maxLength": 255,
            "pattern": "^[A-Za-z0-9 áéíóúÁÉÍÓÚñÑ]+$",
            "description": "Denominación o nombre de la línea. No puede superar los 255 caracteres.",
            "example": "tornillos"
          },
          "stockMinimo": {
            "type": "number",
            "description": "Stock mínimo de la línea. Admite hasta 3 decimales.",
            "example": 5
          },
          "utilizaStockMinimo": {
            "type": "boolean",
            "description": "Indica si la línea utiliza stock mínimo.",
            "example": false
          },
          "observacion": {
            "type": "string",
            "description": "Observaciones varias sobre la línea.",
            "example": ""
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Fecha y hora de creación (generada por el servidor).",
            "example": "2026-09-23T15:00:00.000Z"
          },
          "usuarioCreatedId": {
            "type": "number",
            "minimum": 1,
            "description": "ID del usuario que crea la línea. Debe ser un número entero positivo.",
            "example": 1
          },
          "superLineaId": {
            "type": "number",
            "minimum": 1,
            "description": "ID de la SuperLínea asociada. Debe ser un entero positivo.",
            "example": 1
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "Fecha de eliminación (null si está activa).",
            "example": null
          }
        },
        "required": [
          "denominacion",
          "utilizaStockMinimo",
          "usuarioCreatedId",
          "superLineaId"
        ]
      },
      "SelectOption": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "number",
            "description": "Código identificador de la opción.",
            "example": 1
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de la opción.",
            "example": "tornillos"
          },
          "descripcion": {
            "type": "string",
            "description": "Descripción de la opción.",
            "example": ""
          }
        },
        "required": [
          "codigo",
          "nombre",
          "descripcion"
        ]
      },
      "SuperLineaDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1,
            "description": "ID de la superlínea."
          },
          "denominacion": {
            "type": "string",
            "example": "Herramientas",
            "description": "Denominación o nombre de la superlínea."
          },
          "observacion": {
            "type": "string",
            "example": "Uso industrial",
            "description": "Observaciones varias sobre la superlínea."
          },
          "deletedAt": {
            "type": "string",
            "nullable": true,
            "example": null,
            "description": "Fecha de eliminación (null si está activa)."
          }
        },
        "required": [
          "id",
          "denominacion",
          "deletedAt"
        ]
      },
      "UpdateLineaDto": {
        "type": "object",
        "properties": {
          "denominacion": {
            "type": "string",
            "maxLength": 255,
            "pattern": "^[A-Za-z0-9 áéíóúÁÉÍÓÚñÑ]+$",
            "description": "Nueva denominación o nombre de la línea. No puede superar los 255 caracteres.",
            "example": "tornillos"
          },
          "stockMinimo": {
            "type": "number",
            "description": "Stock mínimo de la línea. Admite hasta 3 decimales.",
            "example": 5
          },
          "utilizaStockMinimo": {
            "type": "boolean",
            "description": "Indica si la línea utiliza stock mínimo.",
            "example": false
          },
          "observacion": {
            "type": "string",
            "description": "Observaciones varias sobre la línea.",
            "example": ""
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "description": "Fecha y hora de creación (generada por el servidor).",
            "example": "2026-09-23T15:00:00.000Z"
          },
          "usuarioUpdatedId": {
            "type": "number",
            "minimum": 1,
            "description": "ID del usuario que actualiza la línea. Debe ser un número entero positivo.",
            "example": 1
          },
          "superLineaId": {
            "type": "number",
            "minimum": 1,
            "description": "ID de la SuperLínea asociada. Debe ser un entero positivo.",
            "example": 1
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "nullable": true,
            "description": "Fecha de eliminación (null si está activa).",
            "example": null
          }
        },
        "required": [
          "usuarioUpdatedId"
        ]
      },
...
    }
  }
}